Added additional delete option to delete properties (not default)

+++ src/com/google/gwt/sample/itProjekt/server/db/PropertyMapper.java
@@ -14,14 +14,14 @@ import com.google.gwt.sample.itProjekt.shared.bo.User;
 
 public class PropertyMapper {
 
-	/** Konstruktor für den PropertyMapper (Singleton) */
+	/** Konstruktor fï¿½r den PropertyMapper (Singleton) */
 	//static weil Singleton. Einzige Instanz dieser Klasse
 	private static PropertyMapper  propertymapper = null;
 	
 	/**
 	 * PropertyMapper.
 	 *
-	 *Falls noch kein ProperyMapper existiert erstellt er ein neuen PropertyMapper und gibt ihn zurück
+	 *Falls noch kein ProperyMapper existiert erstellt er ein neuen PropertyMapper und gibt ihn zurï¿½ck
 	 * 
 	 */
 	public static PropertyMapper propertyMapper() {
@@ -35,7 +35,7 @@ public class PropertyMapper {
 /**
  * FindByID.
  *
- *Findet Property durch eine P_ID und speichert die dazugehörigen Werte (P_ID und type) in einem Property Objekt ab und gibt dieses wieder
+ *Findet Property durch eine P_ID und speichert die dazugehï¿½rigen Werte (P_ID und type) in einem Property Objekt ab und gibt dieses wieder
  *
  */
 public Property findByID(Property property){
@@ -62,9 +62,9 @@ public Property findByID(Property property){
 /**
  * FindAll.
  *
- * Gibt alle Property Objekte zurück welche mit P_ID, type und C_ID befüllt sind
- * Hierfür holen wir die Attribute aus der T_Property Tabelle und speichern diese in einem Property Objekt ab und fügen diese dem Vector hinzu
- * Am Ende geben wir diesen Vector zurück
+ * Gibt alle Property Objekte zurï¿½ck welche mit P_ID, type und C_ID befï¿½llt sind
+ * Hierfï¿½r holen wir die Attribute aus der T_Property Tabelle und speichern diese in einem Property Objekt ab und fï¿½gen diese dem Vector hinzu
+ * Am Ende geben wir diesen Vector zurï¿½ck
  *
  */
 
@@ -93,9 +93,9 @@ Vector<Property> result = new Vector<Property>();
 /**
  * FindAllByCID.
  * 
- * Findet alle P_ID, type und C_ID wo die C_ID der ID des übergebenen Objekte entspricht
- * Befüllt das Property Objekt mit den Attributen und fügt dieses Objekt dem Vector hinzu
- * Gibt ein Vector voller Property Objekte zurück
+ * Findet alle P_ID, type und C_ID wo die C_ID der ID des ï¿½bergebenen Objekte entspricht
+ * Befï¿½llt das Property Objekt mit den Attributen und fï¿½gt dieses Objekt dem Vector hinzu
+ * Gibt ein Vector voller Property Objekte zurï¿½ck
  *
  */
 
@@ -126,11 +126,11 @@ Vector<Property> result = new Vector<Property>();
 /**
  * FindAllDefault.
  *
- * Gibt alle Proeprty Objekte zurück die mit den Default Werten befüllt sind
- * Hierfür sind alle Default Property Werte mit der C_ID von 20000000 verknüpft
- * Alle Attribute die zu dieser C_ID von 20000000 gehören holen wir aus der T_Property Tabelle (P_ID, type und C_ID)
- * und speichern diese in einem Property Objekt ab und fügen diese dem Vector hinzu
- * Am Ende geben wir diesen Vector zurück
+ * Gibt alle Proeprty Objekte zurï¿½ck die mit den Default Werten befï¿½llt sind
+ * Hierfï¿½r sind alle Default Property Werte mit der C_ID von 20000000 verknï¿½pft
+ * Alle Attribute die zu dieser C_ID von 20000000 gehï¿½ren holen wir aus der T_Property Tabelle (P_ID, type und C_ID)
+ * und speichern diese in einem Property Objekt ab und fï¿½gen diese dem Vector hinzu
+ * Am Ende geben wir diesen Vector zurï¿½ck
  *
  */
 
@@ -159,9 +159,9 @@ Vector<Property> result = new Vector<Property>();
 /**
  * Insert.
  *
- * Sucht nach der höchsten P_ID um diese um eins zu erhöhen und als neue P_ID zu nutzen
- * Befüllt T_Property mit P_ID, type und C_ID
- * Eine Property wird zum Schluss zurückgegeben
+ * Sucht nach der hï¿½chsten P_ID um diese um eins zu erhï¿½hen und als neue P_ID zu nutzen
+ * Befï¿½llt T_Property mit P_ID, type und C_ID
+ * Eine Property wird zum Schluss zurï¿½ckgegeben
  *
  */
 
@@ -198,8 +198,8 @@ public Property insert(Property property, Contact contact){
 /**
  * Update.
  *
- * Update von Veränderungen falls sich type ändert
- * Gibt ein Property zurück
+ * Update von Verï¿½nderungen falls sich type ï¿½ndert
+ * Gibt ein Property zurï¿½ck
  * 
  */
 
@@ -223,7 +223,7 @@ return property;}
 /**
  * Delete.
  *
- * Entfernt alles aus T_Property wo die P_ID der ID des übergebenen Objekts entspricht
+ * Entfernt alles aus T_Property wo die P_ID der ID des ï¿½bergebenen Objekts entspricht
  * 
  */
 
@@ -233,7 +233,11 @@ public void delete (Property property){
 				try{
 					
 					Statement stmt = con.createStatement();
+					if(property.getContactID() != 20000000){
+						
 					stmt.executeUpdate("DELETE FROM T_Property WHERE P_ID =" + property.getId());
+					
+					}
 				}
 			
 			catch (SQLException e2){
